update crontab to use latest feature of weekly build/cycling scripts. (#173)
Add an optional commandline argument to specify  an account to use
when submitting compute jobs.

Add an optional commandline argument to specify a suffix to add
to a build directory.

1. put the weekly buils on the scratch filesystem.
2. date the single precision build used for the cycling runs.
3. Update the develop branch of MPAS-Workflow before running the cylc scenarion.

Run the job to create the workflow graphs on casper.
Check every night to see if mpas-bundle should be built.